Delegate consensus launching to a seperate package, to de-clutter the CLI a bit
