+++ pkg/cli/commands.go
@@ -10,7 +10,6 @@ import (
 
 	ristretto "github.com/bwesterb/go-ristretto"
 	"github.com/dusk-network/dusk-blockchain/pkg/config"
-	"github.com/dusk-network/dusk-blockchain/pkg/core/consensus"
 	"github.com/dusk-network/dusk-blockchain/pkg/p2p/wire"
 )
 
@@ -136,16 +135,4 @@ func stringToUint64(s string) (uint64, error) {
 		return 0, err
 	}
 	return (uint64(sInt)), nil
-}
-
-func getStartingRound(blsPubKey []byte, eventBroker wire.EventBroker) uint64 {
-	// Start listening for accepted blocks, regardless of if we found stakes or not
-	acceptedBlockChan, listener := consensus.InitAcceptedBlockUpdate(eventBroker)
-	// Unsubscribe from AcceptedBlock once we're done
-	defer listener.Quit()
-
-	for {
-		blk := <-acceptedBlockChan
-		return blk.Header.Height + 1
-	}
 }
